requirements.txt
SpeechRecognition
pipwin
pyaudio
gTTS
google-generativeai
python-dotenv
streamlit
googletrans==4.0.0-rc1
redis==5.0.0

-e .
